segundo spring

+++ aventon/src/controllers/vehiculoController.php
@@ -7,6 +7,7 @@ public function __construct(){
 public function execute($simpleUrl){
 	$page = $this->evaluarPage($simpleUrl->segment(2)); 
 	SessionHelper::validateSession();
+
 	switch($page){
 		case "home" : $this->home();
 		break;
@@ -33,27 +34,54 @@ public function execute($simpleUrl){
 		break;
 		case "modificar" :  $this->modificar();
 		break;
+		case "nuevo" :  $this->nuevo();
+		break;
+		case "listar" :  $this->listar();
+		break;
+		case "listarPorGuardar": $this->listarPorGuardar();
+		break;
 		default : throw new NotFoundException();
 	}
 }
-
-
-public function init(){
-	$user = SessionHelper::getUser();
-	$vehiculos = VehiculoService::buscarVehiculos($user->getUserId());	
-	$smTemplate = new SMTemplate();
-	$smTemplate->render("verVehiculos",["vehiculos" => $vehiculos,"vehiculo"=>new VehiculoDTO()]);
+public function listarPorGuardar(){
+		SessionHelper::putInSession("volver", true);
+		header('Location: /aventon/vehiculo/listar');
 }
 
+public function listar(){
+	$user = SessionHelper::getUser();
+	$vehiculos = VehiculoService::buscarVehiculos($user->getUserId());	
+	$volver = SessionHelper::getInSession("volver");
+	$mensaje = SessionHelper::getInSession("mensaje");
+	SessionHelper::deleteInSession("volver");
+	SessionHelper::deleteInSession("mensaje");
+	$smTemplate = new SMTemplate();
+	$smTemplate->render("verVehiculos",["vehiculos" => $vehiculos, "volver" => $volver, "mensaje" => $mensaje ]);
+}
+function nuevo(){
+	$ok = SessionHelper::getInSession("OK");
+	SessionHelper::deleteInSession("OK");
+	$smTemplate = new SMTemplate();
+	$smTemplate->render("crearVehiculo",["OK" => $ok,"vehiculo"=>new VehiculoDTO()]);
+}
 public function home(){
-	$this->init();
+	$this->nuevo();
 }
 public function eliminar($idVehiculo){
 	$user = SessionHelper::getUser();
 	$vehiculo = VehiculoService::obtenerVehiculo($idVehiculo,$user->getUserId());
 	if($vehiculo != null){
-		VehiculoService::eliminarVehiculo($idVehiculo);			
-		header('Location: /aventon/vehiculo');
+		//recupero todos los viajes asociados a ese vehiculo
+		$exist = ViajeService::existeViajePendientePorVehiculo($idVehiculo,$user->getUserId());
+		if(!$exist){
+			//Se elimina OK
+			VehiculoService::eliminarVehiculo($idVehiculo);	
+			SessionHelper::putInSession("mensaje", "Se ha eliminado correctamente.");		
+		}else{
+			//retornar un mensaje diciendo que no se puede eliminar
+			SessionHelper::putInSession("mensaje", "No se puede eliminar el vehiculo dado que se encuentra asociado a un viaje sin realizar.");
+		}
+		header('Location: /aventon/vehiculo/listar');
 	}
 	else{
 		throw new NotFoundException();
@@ -73,23 +101,35 @@ public function verVehiculo($idVehiculo){
 }
 public function modificar(){
 	$vehiculo = new VehiculoDTO($_POST);
-	VehiculoService::modificarVehiculo($vehiculo);
-	header('Location: /aventon/vehiculo');
+	$user = SessionHelper::getUser();
+	$exist = ViajeService::existeViajePendientePorVehiculo($vehiculo->getId(),$user->getUserId());
+	if(!$exist){
+		//Se elimina OK
+		VehiculoService::modificarVehiculo($vehiculo);
+		SessionHelper::putInSession("mensaje", "Se ha modificado la información del vehiculo correctamente.");		
+		header('Location: /aventon/vehiculo/listar');	
+	}else{
+		//retornar un mensaje diciendo que no se puede eliminar
+		$mensaje = "No se puede modificar la información el vehiculo dado que se encuentra asociado a un viaje sin realizar.";
+		$smTemplate = new SMTemplate();
+		$smTemplate->render("modificarVehiculo",["vehiculo" => $vehiculo, "mensaje" => $mensaje]);
+	}
 	
 }
 public function guardar(){
 	$vehiculo = new VehiculoDTO($_POST);
 	$user = SessionHelper::getUser();
-	$vehiculoBD = VehiculoService::existPatente($vehiculo->getPatente(), $user->getUserId());
+	$vehiculoBD = VehiculoService::existPatente($vehiculo->getPatente());
 	if($vehiculoBD == null){
 		$vehiculo->setIdUsuario($user->getUserId());
 		VehiculoService::insertVehiculo($vehiculo);  
-		header('Location: /aventon/vehiculo');	
+		SessionHelper::putInSession("OK", true);
+		header('Location: /aventon/vehiculo/nuevo');	
 	}
 	else{//ERROR
 		$vehiculos = VehiculoService::buscarVehiculos($user->getUserId());	
 		$smTemplate = new SMTemplate();
-		$smTemplate->render("verVehiculos",["error"=>true,"vehiculos" => $vehiculos,"vehiculo"=>$vehiculo]);
+		$smTemplate->render("crearVehiculo",["error"=>true,"vehiculos" => $vehiculos,"vehiculo"=>$vehiculo]);
 			
 	}
